Added remove-owner-references function
