+++ remove-ownerowner-references/README.md
@@ -1,26 +1,23 @@
 
-# set-label
+# remove-owner-references
 
 ## Overview
 
-This is a starter code just to understand and setup the workflow for KRM functions.
+The KRM function created in code removes resources with owner references.  
 
-The KRM function created in code adds a label `tier:mysql` to all the resources of kind Deployment.  
 
-## Testing the function :
+## Usage  
 
-## Locally
+To test the function locally:
 
 Convert the KRM resources and FunctionConfig resource to `ResourceList`, and then pipe the ResourceList as stdin to your function
 
 `kpt fn source data | go run main.go`
 
-## Using Docker:
-
+  
+Using Docker
 `export FN_CONTAINER_REGISTRY=<Your  GCR  or  docker  hub>`
-
 `export TAG= <Your  KRM  function  tag>`
-
 `docker build . -t ${FN_CONTAINER_REGISTRY}/${FUNCTION_NAME}:${TAG}`
