+++ remove-ownerowner-references/data/resources.yaml
@@ -16,6 +16,11 @@ metadata: # kpt-merge: /nginx-deployment
   name: nginx-deployment
   labels:
     tier: mysql
+  ownerReferences:
+    - apiVersion: cache.example.com/v1alpha1
+      kind: Memcached
+      name: example-memcached
+      uid: ad834522-d9a5-4841-beac-991ff3798c00
 spec:
   replicas: 3
   selector:
@@ -27,7 +32,7 @@ spec:
         app: nginx
     spec:
       containers:
-      - name: nginx
-        image: nginx:1.14.2
-        ports:
-        - containerPort: 80
+        - name: nginx
+          image: nginx:1.14.2
+          ports:
+            - containerPort: 80
