+++ set-label/Kptfile
@@ -1,0 +1,18 @@
+apiVersion: kpt.dev/v1
+kind: Kptfile
+metadata:
+  name: set-annotation
+upstream:
+  type: git
+  git:
+    repo: https://github.com/GoogleContainerTools/kpt-functions-sdk
+    directory: /go/get-started
+    ref: master
+  updateStrategy: resource-merge
+upstreamLock:
+  type: git
+  git:
+    repo: https://github.com/GoogleContainerTools/kpt-functions-sdk
+    directory: /go/get-started
+    ref: master
+    commit: 78748f08d9977e08b5cf2dd678d04ed22d209a75
